+++ packages/leva/src/styles/stitches.config.ts
@@ -104,24 +104,13 @@ const utils = {
     createStateClass(value, { key: '$active', borderColor: '$accent1', inset: true }),
 }
 
-export const { styled, css, theme, global: _global, keyframes } = createCss({
-  insertionMethod() {
-    let currentCssHead: HTMLHeadElement | null = null
-    let currentCssNode: HTMLElement | null = null
-
-    return (cssText) => {
-      if (typeof document === 'object') {
-        if (!currentCssHead) currentCssHead = document.head || document.documentElement
-        if (!currentCssNode)
-          currentCssNode =
-            document.getElementById('leva__stitches') ||
-            Object.assign(document.createElement('style'), { id: 'leva__stitches' })
-        if (!currentCssNode.parentNode) currentCssHead.append(currentCssNode)
-
-        currentCssNode.textContent = cssText
-      }
-    }
-  },
+export const {
+  styled,
+  css,
+  theme,
+  global: _global,
+  keyframes,
+} = createCss({
   prefix: 'leva',
   theme: getDefaultTheme(),
   utils: {