+++ packages/leva/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Patch Changes
 
+- dd17591: chore: update to stitches 0.2.0
 - d5a5dd2: - Fix: prevent selection on Firefox when closing the Color picker.
   - Fix: add suffix to Number plugin initial value.
   - Feat: add `title` attribute to Labels when there is no `hint` prop.